Adds pygame recipe
As discussed on the Discord, this is a recipe to build apps based on
SDL2-based pygame.

It currently references a RC until we have an official stable
release of pygame based on SDL2 with android support.

Simple examples have been tested by other pygame users (it doesn't just
build on my own machine) but some pygame functionality is still untested,
and some dependencies like freetype, postmidi and libjpeg are currently
not part of the build. It's usable, but not complete.